refactor: move test utils
